first-step refactor, tests not tied in
