+++ atomics/include/desul/atomics.hpp
@@ -1,0 +1,18 @@
+/* 
+Copyright (c) 2019, Lawrence Livermore National Security, LLC
+and DESUL project contributors. See the COPYRIGHT file for details.
+Source: https://github.com/desul/desul
+
+SPDX-License-Identifier: (BSD-3-Clause)
+*/
+
+#ifndef DESUL_ATOMICS_HPP_
+#define DESUL_ATOMICS_HPP_
+
+#include "atomics/Macros.hpp"
+
+#include "atomics/Compare_Exchange.hpp"
+#include "atomics/Generic.hpp"
+#include "atomics/Lock_Array.hpp"
+
+#endif
